+++ helpme/utils/__init__.py
@@ -1,11 +1,3 @@
-from .terminal import (
-    confirm_prompt,
-    choice_prompt,
-    regexp_prompt,
-    run_command,
-    get_installdir,
-    which,
-)
 from .fileio import (
     copyfile,
     get_userhome,
@@ -19,6 +11,14 @@ from .fileio import (
     write_file,
     write_json,
 )
+from .format import envars_to_markdown, format_code_block, generate_identifier_hash
+from .metrics import MetricsCollector
+from .terminal import (
+    confirm_prompt,
+    choice_prompt,
+    regexp_prompt,
+    run_command,
+    get_installdir,
+    which,
+)
 from .settings import get_configfile, generate_keypair, load_keypair
-
-from .format import envars_to_markdown